+++ packages/cli/default/extensions/filters/directus.ts
@@ -4,11 +4,11 @@ import { json } from "./json";
 import { quote, quoted } from "./quote";
 
 import { match } from "ts-pattern";
-import { lower_case, pascal_case, space_case } from "./string_cases";
-import { drop_first } from "./drop_first";
-import { split } from "./split";
-import { regex_replace } from "./regex_replace";
 import type { TemplateContext } from "../../../types/template";
+import { drop_first } from "./drop_first";
+import { regex_replace } from "./regex_replace";
+import { split } from "./split";
+import { lower_case, pascal_case, space_case } from "./string_cases";
 
 export function to_collection_text(
   context: TemplateContext,
@@ -58,25 +58,138 @@ export function to_ts_type(context: TemplateContext, field: Field) {
   let schema = field.type;
   let meta = field.type.raw.meta;
   let nullable = false;
+  let isUnknown = true;
 
-  let db_type = match(field?.type?.database)
+  const _push = types.push;
+  types.push = (...args: string[]): number => {
+    isUnknown = false;
+    return _push.apply(
+      types,
+      args.filter((v) => v != "never"),
+    );
+  };
+
+  let db_type = match(
+    field?.type?.database?.split("(", 2)[0]!.toLowerCase().trim(),
+  )
     .returnType<string | false>()
-    .with("uuid", () => "UUID")
-    .with("json", () => "UUID")
-    .with("text", () => "string")
-    .with("integer", () => "number")
-    .with("decimal", () => "number")
-    .with("numeric", () => "number")
-    .with("bigint", () => "BigInt")
-    .with("boolean", () => "boolean")
-    .with("character varying", () => "string")
-    .with("date", () => "Date")
-    .with("time", () => "Date")
-    .with("time with time zone", () => "Date")
-    .with("time without time zone", () => "Date")
-    .with("timestamp", () => "Date")
-    .with("timestamp with time zone", () => "Date")
-    .with("timestamp without time zone", () => "Date")
+    .with("uuid", () => "Types.UUID")
+    .with("json", () => "Types.JSON")
+    .with("text", () => "Types.String")
+    .with("integer", () => "Types.Integer")
+    .with("decimal", () => "Types.Decimal")
+    .with("numeric", () => "Types.Number")
+    .with("bigint", () => "Types.BigInteger")
+    .with("boolean", () => "Types.Boolean")
+    .with("character varying", () => "Types.String")
+    .with("date", () => "Types.Date")
+    .with("time", () => "Types.DateTime")
+    .with("time with time zone", () => "Types.DateTime")
+    .with("time without time zone", () => "Types.DateTime")
+    .with("timestamp", () => "Types.DateTime")
+    .with("timestamp with time zone", () => "Types.DateTime")
+    .with("timestamp without time zone", () => "Types.DateTime")
+
+    // Shared
+    .with("boolean", () => "Types.Boolean")
+    .with("tinyint", () => "Types.Integer")
+    .with("smallint", () => "Types.Integer")
+    .with("mediumint", () => "Types.Integer")
+    .with("int", () => "Types.Integer")
+    .with("integer", () => "Types.Integer")
+    .with("serial", () => "Types.Integer")
+    .with("bigint", () => "Types.BigInteger")
+    .with("bigserial", () => "Types.BigInteger")
+    .with("clob", () => "Types.Text")
+    .with("tinytext", () => "Types.Text")
+    .with("mediumtext", () => "Types.Text")
+    .with("longtext", () => "Types.Text")
+    .with("text", () => "Types.Text")
+    .with("varchar", () => "Types.String")
+    .with("longvarchar", () => "Types.String")
+    .with("varchar2", () => "Types.String")
+    .with("nvarchar", () => "Types.String")
+    .with("image", () => "Types.Binary")
+    .with("ntext", () => "Types.Text")
+    .with("char", () => "Types.String")
+    .with("date", () => "Types.Date")
+    .with("datetime", () => "Types.DateTime")
+    .with("dateTime", () => "Types.DateTime")
+    .with("timestamp", () => "Types.DateTime")
+    .with("time", () => "Types.DateTime")
+    .with("float", () => "Types.Float")
+    .with("double", () => "Types.Float")
+    .with("double precision", () => "Types.Float")
+    .with("real", () => "Types.Float")
+    .with("decimal", () => "Types.Decimal")
+    .with("numeric", () => "Types.Integer")
+
+    // Geometries
+    .with("geometry", () => "Types.Geometry.Geometry")
+    .with("point", () => "Types.Geometry.Point")
+    .with("linestring", () => "Types.Geometry.LineString")
+    .with("polygon", () => "Types.Geometry.Polygon")
+    .with("multipoint", () => "Types.Geometry.MultiPoint")
+    .with("multilinestring", () => "Types.Geometry.MultiLineString")
+    .with("multipolygon", () => "Types.Geometry.MultiPolygon")
+
+    // MySQL
+    .with("string", () => "Types.Text")
+    .with("year", () => "Types.Integer")
+    .with("blob", () => "Types.Binary")
+    .with("mediumblob", () => "Types.Binary")
+    .with("int unsigned", () => "Types.Integer")
+    .with("tinyint unsigned", () => "Types.Integer")
+    .with("smallint unsigned", () => "Types.Integer")
+    .with("mediumint unsigned", () => "Types.Integer")
+    .with("bigint unsigned", () => "Types.Integer")
+
+    // MS SQL
+    .with("bit", () => "Types.Boolean")
+    .with("smallmoney", () => "Types.Float")
+    .with("money", () => "Types.Float")
+    .with("datetimeoffset", () => "Types.DateTime")
+    .with("datetime2", () => "Types.DateTime")
+    .with("smalldatetime", () => "Types.DateTime")
+    .with("nchar", () => "Types.Text")
+    .with("binary", () => "Types.Binary")
+    .with("varbinary", () => "Types.Binary")
+    .with("uniqueidentifier", () => "Types.UUID")
+
+    // Postgres
+    .with("json", () => "Types.JSON")
+    .with("jsonb", () => "Types.JSON")
+    .with("uuid", () => "Types.UUID")
+    .with("int2", () => "Types.Integer")
+    .with("serial4", () => "Types.Integer")
+    .with("int4", () => "Types.Integer")
+    .with("serial8", () => "Types.Integer")
+    .with("int8", () => "Types.Integer")
+    .with("bool", () => "Types.Boolean")
+    .with("character varying", () => "Types.String")
+    .with("character", () => "Types.String")
+    .with("interval", () => "Types.String")
+    .with("_varchar", () => "Types.String")
+    .with("bpchar", () => "Types.String")
+    .with("timestamptz", () => "Types.DateTime")
+    .with("timestamp with time zone", () => "Types.DateTime")
+    .with("timestamp with local time zone", () => "Types.DateTime")
+    .with("timestamp without time zone", () => "Types.Date")
+    .with("timestamp without local time zone", () => "Types.Date")
+    .with("timetz", () => "Types.DateTime")
+    .with("time with time zone", () => "Types.DateTime")
+    .with("time without time zone", () => "Types.DateTime")
+    .with("float4", () => "Types.Float")
+    .with("float8", () => "Types.Float")
+    .with("citext", () => "Types.Text")
+
+    // Oracle
+    .with("number", () => "Types.Integer")
+    .with("sdo_geometry", () => "Types.Geometry.Geometry")
+
+    // SQLite
+    .with("integerfirst", () => "Types.Integer")
+
     .otherwise(() => false);
 
   if (db_type) {
@@ -86,15 +199,15 @@ export function to_ts_type(context: TemplateContext, field: Field) {
   let json_type: string | false = false;
   if (field.type.is_json) {
     if ("json_schema" in schema) {
-      json_type = '"json_schema"';
+      json_type = "Types.JSONSchema";
     } else {
-      json_type = "any";
+      json_type = "Types.JSON";
     }
   }
 
   switch (meta.interface) {
     case "tags":
-      types.unshift("string[]");
+      types.unshift("Types.String[]");
       break;
     case "select-dropdown":
       let values = (meta?.options?.choices || []).map((v: any) =>
@@ -156,13 +269,13 @@ export function to_ts_type(context: TemplateContext, field: Field) {
   if (types.length <= 0) {
     let schemaStr = json(context, schema);
     let metaStr = json(context, meta);
-    let unknown = `UnknownType<{ schema: ${schemaStr}, meta: ${metaStr} }>`;
+    let unknown = `Types.UnknownType<{ schema: ${schemaStr}, meta: ${metaStr} }>`;
     types.unshift(unknown);
   }
 
   let output = types.join(" | ");
   if (nullable) {
-    output = `Optional<${output}>`;
+    output = `Types.Optional<${output}>`;
   }
 
   return output;